+++ tests/testthat/test-settings.R
@@ -7,9 +7,9 @@ test_that("use_multicore", {
   use_multicore(n_cores = 1, verbose = F)
 })
 
-test_that("use_disk4frames", {
-  expect_output(use_disk4frames(use_disk = TRUE, n_memory_frames = NULL))
+test_that("use_disk", {
+  expect_output(use_disk(frames_to_disk = TRUE, n_memory_frames = NULL))
   expect_is(frames_spatial(m.aligned[move::timestamps(m.aligned) > as.POSIXct("2018-05-15 07:00:00") & move::timestamps(m.aligned) < as.POSIXct("2018-05-15 10:00:00"),],
                            r_grad, r_times, r_type = "gradient", verbose = F, fade_raster = T), "list")
-  use_disk4frames(use_disk = FALSE, verbose = F)
+  use_disk(frames_to_disk = FALSE, verbose = F)
 })